fixes Werror=conversion

+++ src/include/souffle/datastructure/BTree.h
@@ -1859,8 +1859,8 @@ public:
         out << "  Size of leaf node:  " << sizeof(leaf_node) << "\n";
         out << "  Size of Key:        " << sizeof(Key) << "\n";
         out << "  max keys / node:  " << node::maxKeys << "\n";
-        out << "  avg keys / node:  " << (size() / (double)nodes) << "\n";
-        out << "  avg filling rate: " << ((size() / (double)nodes) / node::maxKeys) << "\n";
+        out << "  avg keys / node:  " << ((double)size() / (double)nodes) << "\n";
+        out << "  avg filling rate: " << (((double)size() / (double)nodes) / node::maxKeys) << "\n";
         out << " ---------------------------------\n";
         out << "  insert-hint (hits/misses/total): " << hint_stats.inserts.getHits() << "/"
             << hint_stats.inserts.getMisses() << "/" << hint_stats.inserts.getAccesses() << "\n";
